Fixed rounding

+++ Models/XModel.cs
@@ -5,15 +5,15 @@ namespace AlgorytmGenetyczny.Models
     public class XModel
     {
         [DisplayName("a")]
-        public double RangeBeginning { get; set; }
+        public float RangeBeginning { get; set; }
         [DisplayName("b")]
-        public double RangeEnd { get; set; }
+        public float RangeEnd { get; set; }
         [DisplayName("d")]
         public double Precision { get; set; }
         [DisplayName("N")]
         public int Number { get; set; }
         [DisplayName("xReal")]
-        public double XReal1 { get; set; }
+        public float XReal1 { get; set; }
         [DisplayName("xInt")]
         public int XInt1 { get; set; }
         [DisplayName("xBin")]
@@ -21,9 +21,9 @@ namespace AlgorytmGenetyczny.Models
         [DisplayName("xInt")]
         public int XInt2 { get; set; }
         [DisplayName("xReal")]
-        public double XReal2 { get; set; }
+        public float XReal2 { get; set; }
         [DisplayName("f(x)")]
-        public double FunctionValue { get; set; }
+        public float FunctionValue { get; set; }
 
         public XModel(XModel xModel)
         {
@@ -31,6 +31,6 @@ namespace AlgorytmGenetyczny.Models
             this.RangeEnd = xModel.RangeEnd;
             this.Precision = xModel.Precision;
         }
-        public XModel() {}
+        public XModel() { }
     }
 }
